[WEF-385] 관심종목 api 구현

+++ src/main/java/com/solv/wefin/domain/trading/market/service/MarketService.java
@@ -132,7 +132,7 @@ public class MarketService implements MarketPriceProvider, ExchangeRateProvider 
     @Override
     public BigDecimal getCurrentPrice(String stockCode) {
         PriceResponse response = getPrice(stockCode);
-        return BigDecimal.valueOf(response.currentPrice());
+        return response.currentPrice();
     }
 
     @Override
